- 퀘스트 모든 스테이지 적용 - - 몹 버그 수정 - 여러 버그 수정 - 보석 먹을때 룰렛 어떻게 쓰는지 안내 만들기. - 밸런스 조정 - 룰렛이 돌고 있을때 다른 룰렛 안 돌게하기. - 1-3 공격 해결 - 캐릭터 적 속도 표준화 - 2_1 2_2 리젠 수 낮추기 - 디버그 기능 추가 - 등등 여러 기능들 더 추가함

+++ Lagacy_of_Dragon/stage1_boss.h
@@ -10,7 +10,7 @@ inline bool boss1_move_check = false;
 inline int boss1_x = 750;
 inline int boss1_y = 200;
 inline int s1bossSize = 100;
-inline int boss1_hp = 100;
+inline double boss1_hp = 100;
 inline int s1_boss_vel = 7;
 
 constexpr int b1_hp_x = 250;
